To avoid build  mfgtool-initramfs fail.
Add IMAGE_FSTYPES = "wic.bz2" will trigger to build u-boot-asus-imx or
u-boot-asus-imx-local.

error log:
ERROR: u-boot-asus-imx-local-2022.04-r0 do_deploy_setscene: The recipe
 u-boot-asus-imx-local is trying to install files into a shared area
 when those files already exist. Those files and their manifest location
  are:
  /source/build_imx8mp-blizzard2g_local_mfgtool/tmp/deploy/images/
  imx8mp-blizzard2g/imx-boot-tools/imx8mp-blizzard.dtb
    (matched in manifest-imx8mp_blizzard2g-u-boot-asus-imx-local-mfgtool.deploy)
Please verify which recipe should provide the above files.

The build has stopped, as continuing in this scenario WILL break
things - if not now, possibly in the future (we've seen builds
fail several months later). If the system knew how to recover
from this automatically it would, however there are several
different scenarios which can result in this and we don't
know which one this is. It may be you have switched providers
of something like virtual/kernel (e.g. from linux-yocto to
linux-yocto-dev), in that case you need to execute the clean
task for both recipes and it will resolve this error. It may
be you changed DISTRO_FEATURES from systemd to udev or vice
versa. Cleaning those recipes should again resolve this
error, however switching DISTRO_FEATURES on an existing
build directory is not supported - you should really clean
out tmp and rebuild (reusing sstate should be safe).
It could be the overlapping files detected are harmless
in which case adding them to SSTATE_ALLOW_OVERLAP_FILES
may be the correct solution. It could also be your build
is including two different conflicting versions of thing
(e.g. bluez 4 and bluez 5 and the correct solution for that
would be to resolve the conflict. If in doubt, please ask
on the mailing list, sharing the error and filelist above.
ERROR: u-boot-asus-imx-local-2022.04-r0 do_deploy_setscene:
If the above message is too much, the simpler version is
you're advised to wipe out tmp and rebuild (reusing sstate
is fine). That will likely fix things in most (but not all) cases.

Change-Id: If057fb7256a8411203fb948a598db764d916e992